refactor: inline single-use view hooks
