--- frontend/hooks/useGifts.js
@@ -1,150 +1,0 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useApp } from '../contexts/AppContext';
-import { useToast } from '../contexts/ToastContext';
-import { errorText } from '../lib/helpers';
-import { t } from '../lib/i18n';
-import { announce } from '../lib/announce';
-import { createEmptyGiftForm } from '../lib/gifts';
-import * as api from '../lib/api';
-
-function priceToCents(value) {
-  if (value === '' || value === null || value === undefined) return null;
-  const num = Number(String(value).replace(',', '.'));
-  if (!Number.isFinite(num) || num < 0) return null;
-  return Math.round(num * 100);
-}
-
-export function useGifts() {
-  const { familyId, messages, demoMode, isChild } = useApp();
-  const { success: toastSuccess, error: toastError } = useToast();
-
-  const [gifts, setGifts] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const [statusFilter, setStatusFilter] = useState('');
-  const [recipientFilter, setRecipientFilter] = useState('');
-  const [includeGifted, setIncludeGifted] = useState(false);
-  const [sortOrder, setSortOrder] = useState('created_desc');
-  const [form, setForm] = useState(() => createEmptyGiftForm());
-  const [editingId, setEditingId] = useState(null);
-
-  const loadGifts = useCallback(async (fid = familyId) => {
-    if (!fid || demoMode || isChild) {
-      setGifts([]);
-      return;
-    }
-    setLoading(true);
-    const { ok, data } = await api.apiGetGifts(fid, {
-      status: statusFilter || null,
-      forUserId: recipientFilter ? Number(recipientFilter) : null,
-      includeGifted,
-      sort: sortOrder,
-    });
-    if (ok && data?.items) setGifts(data.items);
-    setLoading(false);
-  }, [familyId, demoMode, isChild, statusFilter, recipientFilter, includeGifted, sortOrder]);
-
-  useEffect(() => {
-    loadGifts();
-  }, [loadGifts]);
-
-  const filteredGifts = useMemo(() => gifts, [gifts]);
-
-  const resetForm = useCallback(() => {
-    setForm(createEmptyGiftForm());
-    setEditingId(null);
-  }, []);
-
-  const populateForm = useCallback((gift) => {
-    setEditingId(gift.id);
-    setForm({
-      title: gift.title || '',
-      description: gift.description || '',
-      url: gift.url || '',
-      for_user_id: gift.for_user_id ? String(gift.for_user_id) : '',
-      for_person_name: gift.for_person_name || '',
-      occasion: gift.occasion || '',
-      occasion_date: gift.occasion_date || '',
-      status: gift.status || 'idea',
-      notes: gift.notes || '',
-      price_eur: gift.current_price_cents != null ? (gift.current_price_cents / 100).toFixed(2) : '',
-    });
-  }, []);
-
-  async function submitGift(e) {
-    e.preventDefault();
-    if (demoMode) {
-      toastError(t(messages, 'module.gifts.demo_blocked'));
-      return false;
-    }
-    const payload = {
-      title: form.title.trim(),
-      description: form.description.trim() || null,
-      url: form.url.trim() || null,
-      for_user_id: form.for_user_id ? Number(form.for_user_id) : null,
-      for_person_name: form.for_person_name.trim() || null,
-      occasion: form.occasion || null,
-      occasion_date: form.occasion_date || null,
-      status: form.status,
-      notes: form.notes.trim() || null,
-      current_price_cents: priceToCents(form.price_eur),
-    };
-
-    if (editingId) {
-      const { ok, data } = await api.apiUpdateGift(editingId, payload);
-      if (!ok) {
-        toastError(errorText(data?.detail, t(messages, 'toast.error'), messages));
-        return false;
-      }
-      toastSuccess(t(messages, 'module.gifts.updated'));
-      announce(t(messages, 'module.gifts.updated'));
-    } else {
-      const { ok, data } = await api.apiCreateGift({ family_id: Number(familyId), ...payload });
-      if (!ok) {
-        toastError(errorText(data?.detail, t(messages, 'toast.error'), messages));
-        return false;
-      }
-      toastSuccess(t(messages, 'module.gifts.created'));
-      announce(t(messages, 'module.gifts.created'));
-    }
-    resetForm();
-    await loadGifts();
-    return true;
-  }
-
-  async function updateStatus(giftId, status) {
-    const { ok, data } = await api.apiUpdateGift(giftId, { status });
-    if (!ok) {
-      toastError(errorText(data?.detail, t(messages, 'toast.error'), messages));
-      return;
-    }
-    await loadGifts();
-  }
-
-  async function deleteGift(giftId) {
-    const { ok, data } = await api.apiDeleteGift(giftId);
-    if (!ok) {
-      toastError(errorText(data?.detail, t(messages, 'toast.error'), messages));
-      return;
-    }
-    toastSuccess(t(messages, 'module.gifts.deleted'));
-    announce(t(messages, 'module.gifts.deleted'));
-    await loadGifts();
-  }
-
-  return {
-    gifts: filteredGifts,
-    loading,
-    statusFilter, setStatusFilter,
-    recipientFilter, setRecipientFilter,
-    includeGifted, setIncludeGifted,
-    sortOrder, setSortOrder,
-    form, setForm,
-    editingId,
-    submitGift,
-    updateStatus,
-    deleteGift,
-    populateForm,
-    resetForm,
-    reload: loadGifts,
-  };
-}
